+++ interpreter/k8s/src/main/scala/io/buoyant/transformer/k8s/LocalNodeTransformerInitializer.scala
@@ -1,7 +1,8 @@
 package io.buoyant.transformer
 package k8s
 
-import io.buoyant.namer.{NameTreeTransformer, TransformerConfig, TransformerInitializer}
+import com.twitter.finagle.Address
+import io.buoyant.namer._
 import java.net.InetAddress
 
 class LocalNodeTransformerInitializer extends TransformerInitializer {
@@ -9,15 +10,28 @@ class LocalNodeTransformerInitializer extends TransformerInitializer {
   override val configId = "io.l5d.k8s.localnode"
 }
 
-class LocalNodeTransformerConfig extends TransformerConfig {
+case class LocalNodeTransformerConfig(hostNetwork: Option[Boolean])
+  extends TransformerConfig {
 
   override def mk(): NameTreeTransformer = {
-    val ip = sys.env.getOrElse(
-      "POD_IP",
-      throw new IllegalArgumentException("POD_IP env variable must be set to the pod's IP")
-    )
-    val local = InetAddress.getByName(ip)
-    new SubnetLocalTransformer(local, Netmask("255.255.255.0"))
+    if (hostNetwork.getOrElse(false)) {
+      val nodeName = sys.env.getOrElse(
+        "NODE_NAME",
+        throw new IllegalArgumentException(
+          "NODE_NAME env variable must be set to the node's name"
+        )
+      )
+      new MetadataFiltertingNameTreeTransformer(NodeNameMetaKey, Some(nodeName))
+    } else {
+      val ip = sys.env.getOrElse(
+        "POD_IP",
+        throw new IllegalArgumentException(
+          "POD_IP env variable must be set to the pod's IP"
+        )
+      )
+      val local = InetAddress.getByName(ip)
+      new SubnetLocalTransformer(local, Netmask("255.255.255.0"))
+    }
   }
 
 }